KARMA-107 Added message on deletion attempt of knowledge area when KAs exist

+++ karma-angular/src/main/webapp/app/entities/space/space-delete-dialog.component.ts
@@ -1,12 +1,15 @@
+import { ArticleService } from '../article';
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { JhiEventManager } from 'ng-jhipster';
+import { JhiEventManager, JhiAlertService } from 'ng-jhipster';
 
 import { Space } from './space.model';
 import { SpacePopupService } from './space-popup.service';
 import { SpaceService } from './space.service';
+import { HttpErrorResponse } from '@angular/common/http';
+import { HttpResponse } from '@angular/common/http';
 
 @Component({
     selector: 'jhi-space-delete-dialog',
@@ -15,11 +18,14 @@ import { SpaceService } from './space.service';
 export class SpaceDeleteDialogComponent {
 
     space: Space;
+    articleCount: number;
 
     constructor(
+        private articleService: ArticleService,
         private spaceService: SpaceService,
         public activeModal: NgbActiveModal,
-        private eventManager: JhiEventManager
+        private eventManager: JhiEventManager,
+        private jhiAlertService: JhiAlertService
     ) {
     }
 
@@ -28,13 +34,30 @@ export class SpaceDeleteDialogComponent {
     }
 
     confirmDelete(id: number) {
-        this.spaceService.delete(id).subscribe((response) => {
-            this.eventManager.broadcast({
+      this.articleService.countBySpace(id).subscribe(
+        (res: HttpResponse<number>) => {
+          this.articleCount = res.body;
+
+          if (this.articleCount === 0) {
+            this.spaceService.delete(id).subscribe((response) => {
+              this.eventManager.broadcast({
                 name: 'spaceListModification',
                 content: 'Deleted an space'
+              });
+              this.activeModal.dismiss(true);
             });
+          } else {
             this.activeModal.dismiss(true);
-        });
+            this.jhiAlertService.error('Must delete associated Articles first', null, null);
+          }
+        },
+        (res: HttpErrorResponse) => this.onError(res.message)
+      );
+
+    }
+
+   private onError(error) {
+        this.jhiAlertService.error(error.message, null, null);
     }
 }
